ENG-7409 rename interface to NeuroIDPublic and revert NeuroIDImpl to NeuroID to reduce changes
