+++ NeuroID/src/main/java/com/neuroid/tracker/callbacks/FragmentCallbacks.kt
@@ -6,7 +6,7 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentManager.FragmentLifecycleCallbacks
-import com.neuroid.tracker.NeuroIDImpl
+import com.neuroid.tracker.NeuroID
 import com.neuroid.tracker.events.RegistrationIdentificationHelper
 import com.neuroid.tracker.events.WINDOW_LOAD
 import com.neuroid.tracker.events.WINDOW_UNLOAD
@@ -14,7 +14,7 @@ import com.neuroid.tracker.utils.NIDLogWrapper
 
 class FragmentCallbacks(
     isChangeOrientation: Boolean,
-    val neuroIDImpl: NeuroIDImpl,
+    val neuroIDImpl: NeuroID,
     val logger: NIDLogWrapper,
     val registrationHelper: RegistrationIdentificationHelper,
 ) : FragmentLifecycleCallbacks() {
@@ -32,11 +32,11 @@ class FragmentCallbacks(
         logger.d(msg = "onFragmentAttached $className")
 
         if (blackListFragments.any { it == className }.not()) {
-            if (NeuroIDImpl.screenName.isEmpty()) {
-                NeuroIDImpl.screenName = "AppInit"
+            if (NeuroID.screenName.isEmpty()) {
+                NeuroID.screenName = "AppInit"
             }
-            if (NeuroIDImpl.screenFragName.isEmpty()) {
-                NeuroIDImpl.screenFragName = className
+            if (NeuroID.screenFragName.isEmpty()) {
+                NeuroID.screenFragName = className
             }
 
             neuroIDImpl.captureEvent(
