+++ NeuroID/src/main/java/com/neuroid/tracker/service/NIDJobServiceManager.kt
@@ -4,7 +4,7 @@ import android.app.ActivityManager
 import android.app.Application
 import android.content.Context
 import androidx.annotation.VisibleForTesting
-import com.neuroid.tracker.NeuroIDImpl
+import com.neuroid.tracker.NeuroID
 import com.neuroid.tracker.callbacks.NIDSensorHelper
 import com.neuroid.tracker.events.CADENCE_READING_ACCEL
 import com.neuroid.tracker.events.LOG
@@ -22,7 +22,7 @@ import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 
 internal class NIDJobServiceManager(
-    private var neuroIDImpl: NeuroIDImpl,
+    private var neuroIDImpl: NeuroID,
     private var dataStore: NIDDataStoreManager,
     private var eventSender: NIDSendingService,
     private var logger: NIDLogWrapper,
@@ -133,7 +133,7 @@ internal class NIDJobServiceManager(
     private fun createSendCadenceServer(): Job {
         return CoroutineScope(dispatcher).launch {
             while (userActive && isActive) {
-                delay(NeuroIDImpl.nidSDKConfig.eventQueueFlushInterval * 1000L)
+                delay(NeuroID.nidSDKConfig.eventQueueFlushInterval * 1000L)
                 sendEventsNotification.send(false)
             }
         }
@@ -144,15 +144,15 @@ internal class NIDJobServiceManager(
      */
     private fun createGyroCadenceServer(): Job {
         return CoroutineScope(dispatcher).launch {
-            while (NeuroIDImpl.isSDKStarted && NeuroIDImpl.nidSDKConfig.gyroAccelCadence) {
-                delay(NeuroIDImpl.nidSDKConfig.gyroAccelCadenceTime)
+            while (NeuroID.isSDKStarted && NeuroID.nidSDKConfig.gyroAccelCadence) {
+                delay(NeuroID.nidSDKConfig.gyroAccelCadenceTime)
 
                 neuroIDImpl.captureEvent(
                     type = CADENCE_READING_ACCEL,
                     attrs =
                         listOf(
                             mapOf(
-                                "interval" to "${NeuroIDImpl.nidSDKConfig.gyroAccelCadenceTime}sec",
+                                "interval" to "${NeuroID.nidSDKConfig.gyroAccelCadenceTime}sec",
                             ),
                         ),
                 )
@@ -205,7 +205,7 @@ internal class NIDJobServiceManager(
             activityManager?.getMemoryInfo(memoryInfo)
         }
 
-        NeuroIDImpl.getInternalInstance()?.lowMemory = memoryInfo.lowMemory
+        NeuroID.getInternalInstance()?.lowMemory = memoryInfo.lowMemory
 
         if (memoryInfo.lowMemory) {
             return NIDEventModel(
