+++ app/src/debug/java/com/sample/neuroid/us/activities/NIDPayloadJsonActivity.kt
@@ -4,9 +4,8 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.neuroid.tracker.NeuroIDImpl
+import com.neuroid.tracker.NeuroID
 import com.neuroid.tracker.utils.NIDLog
-import com.sample.neuroid.us.MyApplicationDemo
 import com.sample.neuroid.us.R
 import com.sample.neuroid.us.databinding.NidActivityJsonPayloadBinding
 
@@ -17,7 +16,7 @@ class NIDPayloadJsonActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setTitle(R.string.nid_custom_events_title_activity)
         binding = DataBindingUtil.setContentView(this, R.layout.nid_activity_json_payload)
-        NeuroIDImpl.getInstance()?.let {
+        NeuroID.getInstance()?.let {
             if (it.isStopped()) {
                 it.start()
             }
