+++ app/src/main/java/com/sample/neuroid/us/MyApplicationDemo.kt
@@ -4,8 +4,9 @@ import android.os.StrictMode
 import android.os.StrictMode.ThreadPolicy
 import android.os.StrictMode.VmPolicy
 import androidx.multidex.MultiDexApplication
-import com.neuroid.tracker.NeuroIDImpl
+import com.neuroid.tracker.NeuroID
 import com.neuroid.tracker.extensions.setVerifyIntegrationHealth
+import com.neuroid.tracker.extensions.startSession
 import com.sample.neuroid.us.domain.config.ConfigHelper
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
@@ -33,15 +34,15 @@ class MyApplicationDemo : MultiDexApplication() {
         )
 
         // tied to form id: form_dream102
-        NeuroIDImpl.Builder(
+        NeuroID.Builder(
             this,
             "key_live_suj4CX90v0un2k1ufGrbItT5",
-            NeuroIDImpl.DEVELOPMENT).build()
-        NeuroIDImpl.getInstance()?.setEnvironmentProduction(true)
-        NeuroIDImpl.getInstance()?.setSiteId(configHelper.formId)
-        NeuroIDImpl.getInstance()?.setVerifyIntegrationHealth(true)
-        NeuroIDImpl.getInstance()?.setUserID(configHelper.userId)
-        NeuroIDImpl.getInstance()?.setRegisteredUserID("ahsdkghasdjkghdklasglasd")
-        NeuroIDImpl.getInstance()?.startSession("testSession")
+            NeuroID.DEVELOPMENT).build()
+        NeuroID.getInstance()?.setEnvironmentProduction(true)
+        NeuroID.getInstance()?.setSiteId(configHelper.formId)
+        NeuroID.getInstance()?.setVerifyIntegrationHealth(true)
+        NeuroID.getInstance()?.setUserID(configHelper.userId)
+        NeuroID.getInstance()?.setRegisteredUserID("ahsdkghasdjkghdklasglasd")
+        NeuroID.getInstance()?.startSession("testSession", true)
     }
 }
