+++ app/src/main/java/com/sample/neuroid/us/activities/NIDCustomEventsActivity.kt
@@ -4,8 +4,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.neuroid.tracker.NeuroIDImpl
-import com.sample.neuroid.us.MyApplicationDemo
+import com.neuroid.tracker.NeuroID
 import com.sample.neuroid.us.R
 import com.sample.neuroid.us.databinding.NidActivityCustomEventsBinding
 
@@ -16,7 +15,7 @@ class NIDCustomEventsActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setTitle(R.string.nid_custom_events_title_activity)
         binding = DataBindingUtil.setContentView(this, R.layout.nid_activity_custom_events)
-        NeuroIDImpl.getInstance()?.let {
+        NeuroID.getInstance()?.let {
             if (it.isStopped()) {
                 it.start()
             }
@@ -28,15 +27,15 @@ class NIDCustomEventsActivity : AppCompatActivity() {
                 showToast()
             }
             buttonSendFormSubmit.setOnClickListener {
-                NeuroIDImpl.getInstance()?.formSubmit()
+                NeuroID.getInstance()?.formSubmit()
                 showToast()
             }
             buttonSendFormSuccess.setOnClickListener {
-                NeuroIDImpl.getInstance()?.formSubmitSuccess()
+                NeuroID.getInstance()?.formSubmitSuccess()
                 showToast()
             }
             buttonSendFormFailure.setOnClickListener {
-                NeuroIDImpl.getInstance()?.formSubmitFailure()
+                NeuroID.getInstance()?.formSubmitFailure()
                 showToast()
             }
         }
