+++ app/src/main/java/com/sample/neuroid/us/activities/sandbox/SandBoxViewModel.kt
@@ -2,8 +2,7 @@ package com.sample.neuroid.us.activities.sandbox
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.neuroid.tracker.NeuroIDImpl
-import com.sample.neuroid.us.MyApplicationDemo
+import com.neuroid.tracker.NeuroID
 import com.sample.neuroid.us.data.network.NIDServices
 import com.sample.neuroid.us.data.network.NetworkInteractor
 import com.sample.neuroid.us.domain.config.ConfigHelper
@@ -37,7 +36,7 @@ class SandBoxViewModel @Inject constructor(
 
     fun checkScore() {
         viewModelScope.launch {
-            NeuroIDImpl.getInstance()?.closeSession()
+            NeuroID.getInstance()?.closeSession()
             delay(2000)
 
             val result = networkInteractor.safeApiCall {
